portfolio/server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
Pillow==10.4.0
